ForTheOneWhoMakesMeSmile
This project is a surprise birthday website created to celebrate a loved one. The site is built using HTML, CSS, Jasascirpt and includes several interactive and emotional elements:

A personalized video message

A fun and engaging mini-game

A heartfelt written greeting

The goal of the project is to deliver a unique, personal, and memorable digital experience. All files are organized to ensure easy deployment and compatibility with standard web browsers.